solved the problem with repetitive code
