+++ src/games/determiningTheLargestDivisor.js
@@ -1,7 +1,6 @@
-const prompt = require('prompt-sync')({ sigint: true });
 const utility = require('./utility');
 
-function processOfFinding() {
+function determiningTheLagerstDivisor() {
   const n1 = utility.randomInteger(1, 100);
   const n2 = utility.randomInteger(1, 100);
   let result;
@@ -13,22 +12,22 @@ function processOfFinding() {
   }
   return result;
 }
-function determiningTheLagerstDivisor() {
-  let nextOne = true;
-  let i = 0;
-  while (i < 3) {
-    if (!nextOne) {
-      break;
-    }
-    const resultGame = processOfFinding();
-    // console.log(resultGame);
-    const gamerAnswer = prompt('Your answer: ');
-    nextOne = utility.audit(resultGame, Number(gamerAnswer));
-    i += 1;
-    if (i === 3 && nextOne) {
-      console.log('You WIN!');
-    }
-  }
-}
+// function determiningTheLagerstDivisor() {
+//   let nextOne = true;
+//   let i = 0;
+//   while (i < 3) {
+//     if (!nextOne) {
+//       break;
+//     }
+//     const resultGame = processOfFinding();
+//     // console.log(resultGame);
+//     const gamerAnswer = prompt('Your answer: ');
+//     nextOne = utility.verificationOfResults(resultGame, Number(gamerAnswer));
+//     i += 1;
+//     if (i === 3 && nextOne) {
+//       console.log('You WIN!');
+//     }
+//   }
+// }
 
 module.exports = { determiningTheLagerstDivisor };
