+++ src/games/progression.js
@@ -1,7 +1,6 @@
-const prompt = require('prompt-sync')({ sigint: true });
 const utility = require('./utility');
 
-function iteration() {
+function progression() {
   let result = '';
   const start = utility.randomInteger(1, 25);
   const difference = utility.randomInteger(1, 6);
@@ -18,21 +17,21 @@ function iteration() {
   return ignorIter * difference + start;
 }
 
-function progression() {
-  let nextOne = true;
-  let i = 0;
-  while (i < 3) {
-    if (!nextOne) {
-      break;
-    }
-    const resultGame = iteration();
-    const gamerAnswer = prompt('Your answer: ');
-    nextOne = utility.audit(Number(resultGame), Number(gamerAnswer));
-    i += 1;
-    if (i === 3 && nextOne) {
-      console.log('You WIN!');
-    }
-  }
-}
+// function progression() {
+//   let nextOne = true;
+//   let i = 0;
+//   while (i < 3) {
+//     if (!nextOne) {
+//       break;
+//     }
+//     const resultGame = iteration();
+//     const gamerAnswer = prompt('Your answer: ');
+//     nextOne = utility.verificationOfResults(Number(resultGame), Number(gamerAnswer));
+//     i += 1;
+//     if (i === 3 && nextOne) {
+//       console.log('You WIN!');
+//     }
+//   }
+// }
 
 module.exports = { progression };
